🌱 Split clusters/ResourceDetailModals.tsx and AddClusterDialog.tsx
Fixes #21799

Split the ResourceDetailModals.tsx barrel file into one file per modal
under resource-modals/ (CPUDetailModal, MemoryDetailModal,
StorageDetailModal, GPUDetailModal) plus a shared.tsx for the common
props/Skeleton/formatMemory helpers; ResourceDetailModals.tsx now
re-exports them so existing import sites keep working unchanged.

Extract useAddClusterForm.ts from AddClusterDialog.tsx: all
form/step/validation state and handlers for the command-line, import,
and connect tabs now live in the hook, leaving AddClusterDialog.tsx as
a thin presentational component (2 hook calls).

No behaviour change.
